+++ src/main/java/com/example/toukapp/dtos/ReservationResponse.java
@@ -14,7 +14,7 @@ public class ReservationResponse {
     private int reservationId;
     private String name;
     private String surname;
-    private int totalPrice;
+    private Float totalPrice;
     private Date date;
     private Date expirationDate;
     private Time expirationTime;
